Added an expression class
The expression class can derive the mathematical
operations for a given line, replacing variables
with arbitrary values given in a map.

It is based on the calculate method.

calculate() is kept since the Expression class contains
more logic.

The purpose of the Expression-class is to enhance the
performance during generation of the LSystem.